requirements.txt
certifi==2022.9.24
click==7.1.2
contourpy==1.0.5
cycler==0.11.0
fonttools==4.25.0
kiwisolver==1.4.2
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
MouseInfo==0.1.3
mss==7.0.1
munkres==1.1.4
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pip==22.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
PyRect==0.2.0
PyScreeze==0.1.28
PySide2==5.15.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytweening==1.0.4
qimage2ndarray==1.9.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
setuptools==65.5.0
shiboken2==5.15.2.1
sip==4.19.13
six==1.16.0
tornado==6.2
wheel==0.37.1
wincertstore==0.2
